refactor: move Mirador stuff to a local mixin
