--- packages/mirador/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@europeana/mirador",
-  "version": "1.119.0",
-  "description": "Mirador for Europeana website IIIF presentation",
-  "license": "EUPL-1.2",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "main": "src/index.js",
-  "scripts": {
-    "build": "npx webpack --config webpack.config.cjs"
-  }
-}
