--- packages/mirador/webpack.config.cjs
@@ -1,10 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  devtool: 'source-map',
-  entry: './src/index.js',
-  output: {
-    filename: 'europeana-mirador.js',
-    path: path.resolve(__dirname, 'dist')
-  }
-};
