+++ packages/portal/Dockerfile
@@ -12,7 +12,6 @@ RUN mkdir -p packages/portal
 
 COPY package*.json ./
 COPY packages/portal/package*.json ./packages/portal/
-COPY packages/mirador ./packages/mirador
 COPY packages/style ./packages/style
 COPY packages/vue-session-id ./packages/vue-session-id
 
@@ -23,7 +22,7 @@ COPY packages/portal/*.js ./packages/portal/
 COPY packages/portal/src ./packages/portal/src
 
 RUN npm run build
-RUN rm -rf node_modules packages/*/node_modules packages/mirador packages/style
+RUN rm -rf node_modules packages/*/node_modules packages/style
 RUN rm -rf packages/portal/src/lang
 RUN npm -w packages/portal ci --omit dev --omit peer
