+++ packages/portal/package.json
@@ -55,7 +55,6 @@
     "weighted-random": "^0.1.0"
   },
   "devDependencies": {
-    "@europeana/mirador": "^1.119.0",
     "@europeana/style": "^1.119.0",
     "@lokalise/node-api": "^8.0.0",
     "@vue-a11y/announcer": "^2.1.0",
