+++ packages/styleguide/Dockerfile
@@ -13,7 +13,6 @@ RUN mkdir -p packages/styleguide packages/portal
 COPY package*.json ./
 COPY packages/portal/package.json ./packages/portal/
 COPY packages/style ./packages/style
-COPY packages/mirador ./packages/mirador
 COPY packages/styleguide/package.json ./packages/styleguide/
 
 RUN npm ci -w packages/portal -w packages/styleguide
